requirements.txt
python-decouple
aiogram
APScheduler